okd_sync/requirements.txt
requests>=2.28.0
psycopg2-binary>=2.9.3
python-dotenv>=0.20.0
boto3>=1.24.0
urllib3>=1.26.0
sqlalchemy>=1.4.0
sqlalchemy-utils>=0.38.0
Pillow>=9.0.0